# docs/homework_openapi.yaml
components:
  schemas:
    Course:
      properties:
        code:
          example: SE2024
          type: string
        created_at:
          example: '2024-01-01 10:00:00'
          format: date-time
          type: string
        description:
          example: 数字特效课程介绍
          type: string
        id:
          example: 1
          type: integer
        is_active:
          example: true
          type: boolean
        name:
          example: 数字特效
          type: string
        teacher_id:
          example: 10
          type: integer
        type: object
    ErrorResponse:
      properties:
        message:
          example: 操作失败
          type: string
        success:
          example: false
          type: boolean
      type: object
...
    SuccessResponse:
      properties:
        data:
          type: object
        message:
          example: 操作成功
          type: string
        success:
          example: true
          type: boolean
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 'JWT Token认证，需要在请求头中携带 Authorization: Bearer <token>'
      scheme: bearer
      type: http
info:
  description: 作业库系统后端接口文档，包含学生端和教师端的所有作业相关功能
  title: 作业库系统API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /api/v2/courses/:
    get:
      description: 教师获取自己的课程列表（仅教师）
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/Course'
                      type: array
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取课程列表
      tags:
      - 课程管理
    post:
      description: 教师创建课程（仅教师）
      requestBody:
        content:
          application/json:
            schema:
              properties:
                code:
                  description: 课程代码
                  example: SE2024
                  type: string
                description:
                  description: 课程描述
                  example: 数字特效课程介绍
                  type: string
                name:
                  description: 课程名称
                  example: 数字特效
                  type: string
              required:
              - name
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      properties:
                        id:
                          example: 1
                          type: integer
                      type: object
                  type: object
          description: 创建成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 创建课程
      tags:
      - 课程管理
  /api/v2/courses/{course_id}:
    delete:
      description: 删除课程（仅教师）
      parameters:
      - description: 课程ID
        in: path
        name: course_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 删除成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 课程不存在
      security:
      - bearerAuth: []
      summary: 删除课程
      tags:
      - 课程管理
    get:
      description: 获取单个课程的详细信息
      parameters:
      - description: 课程ID
        in: path
        name: course_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      $ref: '#/components/schemas/Course'
                  type: object
          description: 获取成功
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 课程不存在
      security:
      - bearerAuth: []
      summary: 获取课程详情
      tags:
      - 课程管理
    put:
      description: 更新课程信息（仅教师）
      parameters:
      - description: 课程ID
        in: path
        name: course_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                code:
                  example: SE2024
                  type: string
                description:
                  example: 数字特效课程介绍
                  type: string
                is_active:
                  example: true
                  type: boolean
                name:
                  example: 数字特效
                  type: string
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 更新成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 课程不存在
      security:
      - bearerAuth: []
      summary: 更新课程
      tags:
      - 课程管理
...
servers:
- description: Local development server
  url: http://localhost:5000
- description: Local network server
  url: http://192.168.8.190:5000
...